Extract account_sets tests to dedicated bootstrap_account_sets.cpp
