+++ nano/core_test/CMakeLists.txt
@@ -14,6 +14,7 @@ add_executable(
   block_processor.cpp
   block_rebroadcaster.cpp
   bootstrap.cpp
+  bootstrap_account_sets.cpp
   bootstrap_frontier_scan.cpp
   bootstrap_server.cpp
   bounded_dfs.cpp
